+++ src/ScottPlot4/ScottPlot.Cookbook/Recipes/Plottable/ScatterList.cs
@@ -98,14 +98,14 @@ namespace ScottPlot.Cookbook.Recipes.Plottable
         }
     }
 
-    public class ScatterListSmooth : IRecipe
+    public class ScatterSmooth : IRecipe
     {
         public ICategory Category => new Categories.PlotTypes.Scatter();
-        public string ID => "scatterList_smooth";
+        public string ID => "scatter_smooth";
         public string Title => "Scatter Plot with Smooth Lines";
         public string Description =>
-            "Set the Smoothness property to a value greater than 1 to draw smooth " +
-            "lines connecting data points. This is similar to the default behavior in Excel.";
+            "Lines drawn between scatter plot points are typically connected with straight lines, " +
+            "but the Smooth property can be enabled to connect points with curves instead.";
 
         public void ExecuteRecipe(Plot plt)
         {
@@ -113,9 +113,8 @@ namespace ScottPlot.Cookbook.Recipes.Plottable
             double[] xs = DataGen.RandomWalk(rand, 20);
             double[] ys = DataGen.RandomWalk(rand, 20);
 
-            var sp = plt.AddScatterList();
-            sp.AddRange(xs, ys);
-            sp.Smoothness = 10;
+            var sp = plt.AddScatter(xs, ys);
+            sp.Smooth = true;
         }
     }
 }